+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.1.1] - 2024-07-14
+
+### Changed
+
+- Updated GOV.UK Frontend to [v5.4.1](https://github.com/alphagov/govuk-frontend/releases/tag/v5.4.1)
+
 ## [1.1.0] - 2024-06-28
 
 ### Added